# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TiendaPc API",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener todas las categorías",
                "operationId": "eabcff1bfb671e6a725645ac6c3dc828",
                "responses": {
                    "200": {
                        "description": "Lista de categorías",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "description": "Nombre de la categoría",
                                                "type": "string"
                                            },
                                            "slug": {
                                                "description": "Slug de la categoría",
                                                "type": "string"
                                            },
                                            "description": {
                                                "description": "Descripción de la categoría",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar todos los productos",
                "operationId": "3e4894c6b444cc2de5e23737d1e4d6c2",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "433d1eb377353a163ba249878346670e",
                "requestBody": {
                    "description": "Datos para crear un producto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto X"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "99.99"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Una descripción opcional"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado con éxito"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto por su ID",
                "operationId": "32d799058b8d35f8deeb4cec51905f2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto obtenido"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "8161b245f9294b1179db922deeea15d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos para actualizar el producto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto Modificado"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "149.99"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nueva descripción"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "258b3511b4c86092cdcccd63375b5b38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión y obtener un token de acceso",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "description": "Credenciales de inicio de sesión",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "description": "Datos para registrar un usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Categorías",
            "description": "Categorías"
        },
        {
            "name": "Productos",
            "description": "Productos"
        },
        {
            "name": "Autenticación",
            "description": "Autenticación"
        }
    ]
}